// Qanat.Web/src/app/shared/generated/model/parcel-minimal-dto.ts
export class ParcelMinimalDto { 
    ParcelID?: number;
    GeographyID?: number;
    WaterAccountID?: number | null;
    ParcelNumber?: string | null;
    ParcelArea?: number;
    ParcelStatusID?: number;
    OwnerAddress?: string | null;
    OwnerName?: string | null;
    WaterAccount?: WaterAccountDisplayDto;
    ParcelStatus?: ParcelStatusSimpleDto;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}